refactor(migrate_app_js#auto): extract a plugin/helper for auto transforms
