+++ src/templates/legacy_app.ejs
@@ -25,7 +25,7 @@ const wrapZafClient = async (client, apiPath, ...rest) => {
 };
 <% }; %>
 
-<% if (helpers.menus) { %>
+<% if (helpers.uiWidgets) { %>
 import ZendeskMenus from "zendesk_menus";
 const {
   Menu,
